Fix compatibility with Elementor
When embedding a contact form on an Elementor page, _mf_captcha_code is set multiple times, but only one captcha image is created on the first run.

This results in an unsolvable captcha since the last stored code doesn't neccessarily match the captcha image.

The exact reason for this behaviour is unknown.

As a workaround, we make sure that generated captcha codes are appended to each other in the session.
We then check against all of those in the mf_submission_validation hook.