Display captcha notice beneath the captcha
We use a hidden field here that is displayed with CSS when the captcha is not solved.
I tried using the internal validation function, but it doesn't work correctly with custom fields in the footer.

It expects your field names to be in the format mform_num_mfield_num, when setting this for fields in the footer
it breaks and returns an empty validation message.

I did not research this further, hence this workaround.